requirements.txt
confusable-homoglyphs==3.2.0
Django==2.2.7
django-auth-ldap==2.0.0
django-autoslug==1.9.6
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-js-asset==1.0.0
django-registration==3.0.1
django-taggit==1.1.0
django-widget-tweaks==1.4.5
mysqlclient==1.4.6
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-ldap==3.2.0
pytz==2019.3
sqlparse==0.3.0
Unidecode==1.1.1
views==0.3
